Solved case 298244: application.getwindowHeight() returns 0 in(D)WC in 5.2. Merge from 5.2.
The IE7 bug Gabriel created a workaround for resulted in window/form size not being correctly sent to server. In order to avoid the IE7 bug (trouble when using both height&width 100%) and still have the sizes, added 2 divs instead of 1 - one for height (with width 0) and one for width (with height 0) - as this doesn't seem to trigger the bug.

Also solved another problem where the WC tableview would enter an infinite loop of re-rendering - lots of traffic + mouse flickers and not useable when anchoring was off.
Case: 298244