The bug of Servoy show different errors when you use WC and SC.
Fixed a problem when the value = null (but the list didnt have a empty value slot and would generate one, but real value would be "" this is bad because real values most of the time are numbes so no that maps to null)

Combobox now doesnt set changed to true when it got a record fill. (because that would cause constant updates and you cant jump from combo to combo in a tableview because of the constant pushing of new select inputs)

if the value is not in the list, make sure that the generated empty value is selected so that that will not generate a select event when pressed on.
Case: 307796