TypeAhead is slow, fast typing ended by enter results: no value
make sure that if the enter key is used (when showing or not showing the popup) the list is filled first and the value is tried to be mapped to a list value. Only do this when there is a task waiting.
In the web we just try to map on a value by looking at if it starts with the typed string (as a last resort)

Case: 355206