Fixed SVY-1136 (also SVY-1117 as duplicate) Webclient: Forms opened in a modal dialog from within another modal dialog do not display later when shown from the main form.
Problem was that even if parent window was not specified in JS, we set one the first time a JSWindow was shown, thus that JSWindow tried to use it whenever it was later shown. And in WC this caused the bug because we didn't check that the parent is still visible.

Case: SVY-1136