servoy_war_export phase 1
Settings are saved to a [user-dir]/.servoy/server/servoy-server.properties when something is modified from the server/admin page.
Which settings file is used is listed on the admin page.

Disabled plugins/beans/lafs upload for now, those should also be uploaded to that dir, but then when updating a war the 'old' plugins remain, so then the war file should win again.. We could also save in the war but then also the special properties file need to be altered because there is where the dates are.

Case: SVY-32