Only get the initialized tables in the gettable() call, better fix for :
performance improvement (one load of solution could trigger almost 300 calls to getTable() for a specific solution)
added getTables() to the remove call, and the server proxy will now ask for all the tables when 1 table is not found. So that from 1 server that is used all the tables are get in 1 go.