Fix for SVY-1308: Deadlock
Removed table lock usage for hidden table flag.
It will rely on the server list sync only. It's not actually a column info related attribute even if it is stored in the same file; it's loaded when table list is loaded.
Also corrected a bad return comment for jsdoc.
Case: SVY-1308