Fixed SVY-1671 : Wrong selected index after databroadcast when deleting rows (6.1 build failing).
Identified 2 problems that were leading to problems with selection and size on this particular jsunit test.
One was a threading issue that resulted in a delete range even being fired much after the actual data was updated.
The other was a very unlikely situation that could occur when one client inserted records and another was refreshing the foundset (those records would be shown twice in the latter).
Case: SVY-1671