Servoy 6.0.6 very slow in building
Various fixes to speed up the creation of FlattenedForms
Cache forms on its id also instead of only by name
Make sure that FF don't fire a change event and fill the extends form on construction. Do this lazy
When filling the FF, look up the script methods or variables on a cache set, not by getScriptMethod(name) call on itself because that is just a full scan.
Case: SVY-1864