Issues and.or possible improvements to UI updates by wicket in the Web Client
2d and 3th point fixes.
make sure that the RecordItemModel always gets the actual component (WebDataCheckbox) and not its wrapper in the lastRenderedValue so that the check later on maps on the same. Now RecordItemmodel always gets the actual component not the inner field. So also the IResolver should be implemented by that.
Before updating the lastRenderedValues the resolver should be asked so that the actual value is stored not just "true" or "false"

other small fix: if the realvalue is already a boolean just return that directly in resolveDisplayValue
Case: SVY-1200