Issue wih onRecordEditStop when using GlobalMethodConverter
by default if looseFocus is false (most of the time when we want to execute a function) we don't call onRecordEditStop

table or lists didn't (try) to save anymore on row selection.

so on record edit stop was almost not called anymore especially because we don't do it anymore on looseFocus = false. (and that was what the current selection event did do)
Case: SVY-2022