Issue with sample solution servoy_sample_pdf_forms
we disabled always all components, but if a component was disabled and the boolean is true
then we did skip it. Problem is that all components are also none servoy components but plain wicket components (like listview items and so on)
So we did call setEnabled(false) on them once but then never true again. Now we only skip the switch from false->true when it is really a IComponent
Case: SVY-2234