'Listener invoked from non event dispatch thread' on sample solutions: _start_here and sample_bug_db
made the log a tracing log, this class is created especially for this reason, that it will fire the events to the dispatch thread, if there are problems because of this then we should look at the cause. not trying to fix it way further up. Foundsets should be changeable in other threads then the event thread.
Case: SVY-2232