Severe degradation in performance when editing records (in a found set) when these records appear on a form at that time within a client
the latest java 6 releases and java 7 have a changed repaint trigger behavior in the RepaintManager, They now have there own runnable pointing to its instance instead of going through the Swing utilities that would result in the actual RepaintManager to be used.

So now the paintDirtyComponents doesn't go through our class but directly to the wrapper.

Changed this so that we don't wrap it anymore but really are the repaintmanager itselfs (super calls instead of delegates)
Case: SVY-2380