Fixed latest problem found during validation after fixing SVY-2619 : Import problem: Forms are not updated after import.
When importing solutions that were exported with older versions of Servoy (that did not support multiple scope names) you would end up with a file "null.js" instead of "globals.js". This didn't happen before as that property was set when creating a new script child to the solution but because of the initial bug it was not removed so it remained "globals". After the initial fix, because the import does not contain any scope name property, that scopeName was cleared => null.
Case: SVY-2619