Cannot navigate to tabpannel tabs, in WC-chrome or firefox. If you click them nothing happens.
don't generate a empty string background-image:url because that results in a query of the normal webpage. This will result in a full rendering of the complete webpage at the server again that is not used anywhere. This resulted in an out of sync between the actual content in the browser and server.
(besides being a huge waste of time and bandwidth) 
Case: SVY-2620