getServerTimeStamp() should convert time if use.client.timezone is true.
application.getTimeStamp() will now always return the actual current time of the client, so in SmartClient this is just new Date() but in the Web it will always convert the date to the actual time of the current locale of the browser. (it doesn't matter what the use client time zone is in this case)

application.getServerTimeStamp() will always return the actual current time of a server, but it will convert it depending on the client and the use client timezone setting:

smartclient: useclienttimezone is true, then it will convert so that the time is compensated so it will display at the smart client the time of the server
smartclient: useclienttimezone is false, then it will do no conversion because the user will see the server time but converted to its own timezone (but the time of the server is getting into the datbase)

webclient: useclienttimezone is true, will just return new Date() so that it will display the server time in the browser even if the browser has a different locale,
webclient: useclienttimezone is false, will act the same as getTimeStamp() so that the client sees the time of the server in its own time, so it is converted to the local time of the browser.

Case: SVY-2726