scopes intialise in random order, causing runtime errors because of references to not yet available scopes
scopes are now created on demand, except the existing "globals" scope we always had, that one is still created at startup.

recursive calls are supported but when a recursive calls hits a scope that was just initializing its variables not all of its variables are there and initalized yet. But all functions can be called..

This change does result that a developer can't depend on a scope to be initialized up front at start.
Case: SVY-3173