'Error calling server' error when using servoy login solution
invoke laters are now done after the current one finishes and goes back the the http thread. Then the http thread calls addEvent again on those that where collected by the script thread for that main event.

Fixed also the redirect to the solution loader if the current mainpage is in a dialog. Make sure that there are no pagemaps in it of that dialog

Case: SVY-3803,SVY-3749