'Error calling server' error when using servoy login solution
cleanup and some more fixes that waiting events are also set to be executed when a suspend happens.
runInBackground can't do this because that would completely kill the runn in background feature. (because waiting events are pushed and waited for on the http thread but the run in background is just running).

Case: SVY-3803,SVY-3749