[WC] Hanged completely with "Loading..." or "Page expired" when using 'servoy.webclient.startscriptthread=true'
when events are suspended then there creator http thread will be moving on. LAter when the event is resumed again another http thread should be waiting for that event to full execute (or suspend again)
For this if a resume happens i add a invokeLater event to the client of the event that does the resume, this will be picked up by the current http thread and that thread will wait for that dummy event to be executed, this will only happen if the resumed event is fully finished or is suspened again.

Also the resumed event should have the right http thread again that is now waiting for itself, so that it can move the pages to itself (and visa versa)
Case: SVY-3895