Developer deadlock
made the isConnected check through the socket.write a bit more lazy and done in a thread.
Now it could be that isConnected lags behind a bit if it says true but it is false. But this shouldn't be to much of a problem. the ui of eclipse calls this quite often so it should return as fast as possible
Case: SVY-3875