WebClient: UI not in sync with reported state in debugger
a WebCellBasedView should set the validation state as soon as possible on the elements and not "later" when it renders. Because that makes the state between scripting different

controller.search()
elelement.txt.editable = false

If txt was editable when doing find then this should result in a none editbale element. But because search() -> validationEnabled = true is done later on the actual element then editable = false. Then the wasEditable of the setValidation() method is overwritting the actual script call


revert of some test code.
Case: SVY-4310