ValueList formatting a integer field to display
LookupValuelist didn't use the field display format if there where realvalues. Now when it has realvalues the lookup valuelists will try to get them from the actual columns of the table (default format) 

The same is being done for DB or Related valuelists those don't have a field display format at all but if they have realValues then the actual columns are being looked at to see if the default format string can be get from there.

in the future we the valuelist should have a format per column (so max 3) so that you can specify it in the valuelist itself.
Case: SVY-4499