java.lang.NullPointerException (no idea what's causing this)
this has to be a double click or fast click on the shutdown client on the server somehow Because it seems to me that 2 shutdowns are almost at the same time active. And because of that the first closes stuff for the second which is then giving those exceptions.
Case: SVY-4585