JSField.text returns always null in Servoy 7.2
this was because JSField.class is not an return type anymore of anything (for that we have that JSFieldWithConstants.class)
And because of that ScriptObjectRegistry.getJavaMembers() was not called on it so there was not an InstanceJavaMembers object created, but it did fall back to the default JavaMembers.
I fixed it in a more general way in the WrapFactory to make sure that all java objects/classes that must be wrapped by the wrapfactory are going through the getJavaMembers call so that any class that should be an InstanceJavaMembers is now always wrapped through that.

Case: SVY-4959