Partial fix for SVY-5277 : Setting dataprovider with name 'valid_to', type 'DATETIME' with value of wrong type '2013-12-31 00:00:00.0'
This is only a partial fix for the whole problem - relying on the fact that the bug gets triggered here by a button push => prepareForSave(true) is called anyway. But a more robust approach is needed.
It will fix the problem in this case, but it can still happen with scripts that are called in other ways, without pushing the UI edits first (for example focus events, schedulers...). I will create a separate case for that.
Case: SVY-5277