For SVY-5263 : Css stylesheets/js files added using WebClientUtils plugin lost.
Global references are now ordered (as they were added). Also the & is quoted correctly in case of media usage in <link or <script tags (before, as a result of the first patch for this same case - in IE it wouldn't load things correctly from an AJAX response; other browsers did complain but they treated the case well anyway).