add solution model support in ngclient
created a global var controllerProvider where controllers can be added and updated on.

ignoreChanges change: now that is only set when the bean data is copied, not when events are called (those can set stuff that should be send to the server again)

The event is suddenly a jquery? object instead the mouse or key browser event object. Unwrap it so that we can get the right info from it.

added method to update the controller and template urls for a given form. Added methods to get form urls now always through methods so that the actual url can change.
Case: SVY-6234