handlers (events) that are design time, should be linked in the link fase
added svy click, doubleclick, rightclick, focusgained and focuslost directives

implemented first 3 for label and button and focus related is added to the textfield

added also test for this to test if they where attached. 

Also added test for the utils directly to also see if it doesn't attach when it is not there.
Case: SVY-6034