[NG] Solution model altered Media
media url is now also adding client uuid and lastmodified timestamp if the media url is part of the copy solution.
FormElements are not cached any more if solution model is used (because media can be altered but the element using that media is not altered at all)
FlattenedSolution now has the time stamp based from solution copy or the normal solution. The solution copy will be updated when a persist is taken for modification
Case: SVY-6357