servoy sample crm: ng-grid shows wrong or no data
removed DEFAULT_PAGE_SIZE. There is no really a default,  it should always be generated (always the first time). Als remember pageSize in the model if we visit the page again. 
On the server try to keep the currentPage if the foundset is the same as the last one that was set (so that you stay on the same page if you toggle between forms)
watch currentPage and pageSize independently and only send that property over so that not constantly the currentPage is toggled between server and client.

added the tmp store of rows again, doesn't work perfect because all forms do already a request data at the moment they are created.  A tableview shouldn't really do that. The first call should be the pagesize calculation.
The first time there are always 2 request (the default and the pagesize) the second time its only the default now because the pagesize is already correct.
Case: SVY-6578