SVY-6615 add server side api support
If there is a serverside script then that is compiled to a script and
then executed against a context object that already has "$scope",
"$scope.model" and "$scope.api" 
those objects then need to be filled in just like on a client.
model will have the model of the webcomponent that is also on the
client.