SVY-7181 Spike: Performance in sample_crm
services "state" objects now are there own angular $scope object (so not
a plain js objects)
where you can call $watch on.
The server side synced state is in a "model" property on that service
scope object.
When an api call or a model sync happens then $digest will be called on
the service scope object