SVY-6923 Complete war export wih solution SVY-6444 i18n keys should be exported in WAR file
everything except the solution stuff it self is not exported as a
solution.servoy in the WEB-INF dir.
This one is then imported the first time, datamodel changes, metadata,
i18n should all be changed/imported into the database.

Only groups are not done yet, because those are part of /the soluton.
(in the export and import)