For SVY-7446 : Finish (to) tagstring refactor property type
Now it is possible for properties to dynamically change at runtime their
record related status (which makes them sent either in model or in
modelViewport in case of portals, listviews, ...).

So if you change through rhino the text property of a label it might
change the way it gets sent to client if it's record dependency state
changes. (no more record linked tags or starts to have some)