SVY-7774 Loss of precision storing database value
getAsRightType should not always convert the NUMBER case to a Double if
the long can't fit into the Double then a BigDecimal should be used.

equalsObject should try to always use BigDecimals to compare if 1 of
them is a bigdecimal, else the conversion to a double could result in
loss of precission an the numbers are then the same.