For SVY-8050 : Adding records in the middle of portal viewport breaks component APIs
Also fixed calling API on a not-yet-created cell. For example if you
have many records in foundset and only 10 are loaded in client - table
view, then on server you select/scroll to bottom and call API, that
needs scroll in client / request additional rows, and execute API after
the scroll is done...

An ui-grid update was needed as well.