SVY-8463 .dividerLocation; doesn't return the right value SVY-8479 Divider location issue SVY-8480 Splitpane divider not shown and overlapping the right side form
fixed various issues, now the resizeWeight is always done compared to
the size that it was before the last calculation. 
made sure that divLocation is always a number
show now the size (normally or when dragging)
initalize the first time only when split1,split2 and the handler are all
there.