SVY-8676 search results are displaying wrong (MINOX)
improved the code that makes the loading of the data a bit lazy.
(because then the full page is rendered a bit better first before the
full tableview is rendered)
now if foundset do change in the mean time, the timeouts are canceled so
only the last foundset is assigned.