Fixed 3 problems noticed while working on SVY-8780 : Event "onShow" is not triggered (MINOX)
(these scenarios are not using recreateUI at all)
So 'main' form has a tabpanel with 2 tabs: 'child1' form and 'child2'
form (both showing related data).

Child2 form always returns false in it's onHide handler.
Here are the  scenarios:

1. main is shown; child1 is shown by default first; click on child2;
click on a button that changes main form completely to some other form;
child2 onHide executes and returns false. But main form is switched
anyway. Problem is that in NG, onHide of child2 is called twice.
2. same as above; but then go back to main form in main window; the main
form's onShow is called but in SC and WC child2's onShow is not called
although it becomes visible... Inconsistency is that in NG, child1's
onShow is called although child1 is not even visible...
---
3. Also fixed a bug where when you have for example 2 child tabs, first
related and second unrelated, then you click on second tab and switch
main form, the second tab would not have the onHide handler called.