Related to SVY-8780 : Event "onShow" is not triggered (MINOX)
Fixed another problem that I noticed when testing this case (not related
to the case itself): when one would call recreateUI on a form that
wasn't modified with SM (or it was modified and already sent to client
once, but then not modified again before calling recreateUI) then the
form URL would not change which means on the client it would not really
be recreated - but on server it would recreate it. Then it got into an
inconsistent state and the client was no longer usable.