Fixed the z-index part of SVY-8928. (centering already works - it was fixed recently)
The idea here is that we create a stacking context with low global
z-index for main form and main navigator form so that their content
doesn't interact with z-indexes of modal windows, popups and so on (they
are independent of each other). So now even though bootstrap css defines
the navigator custom component inside the main form with 1030, it will
still show below the form-in-dialog which has 900, because the 1030 is
relative to the stacking context with z-index 1.