Related to SVY-9895 : Inherited event arguments of an element disapear when overriding some property
Fixed a problem where it wouldn't get rid of inherited method arguments
when you selected some other method completely in the child component.

There is still one more problem to fix: reset to default of handler from
properties view doesn't clear the arguments as well - but this will only
be fixed in 8.1 as the code in that area for 7.4 would require too many
changes to be backported to be able to fix this.

Conflicts:
	servoy_shared/src/com/servoy/j2db/persistence/AbstractBase.java