SVY-10153 html template push in toTemplateJSONValue of FormComponentPropertyType (share the same template, but it could be different per form component..)
the template of a form component now has only the simple names of the
referened form, this way the template is very common and shareable over
many form components, the main form will push all the unique form
component templates to a cache that cache will if it is the first time
precompile the template. The form component implementation only has to
get now the elements and append it to itself.