SVY-10152 FormComponent: Method with arguments (designer and solution model)
customProperties are part of the json but they should be merged into the
current one. So we don't throw away an already set custom property.

If having customProperties directly inside the json is a problem then
there are 2 ways to solve it, in the
json we store it as svy_customProperties or we store all the spec
properties in the "json" indirection property and then we have json and
customProperties as top level