For SVY-10186 SPIKE: improve lazy loading of component's controller
Fixed a bug where a property had design value but it was not present (no
changes) in full to json that we now merge into the template as well.
And as conversions were not merged on top of design but replaced design
conveersions completely, you would end up with a design value that was
being applied without it's needed conversion info.