SVY-10904 Foundset refresh
this was a regression of SVY-9511
an extra method was added but the sub classes override the one that was
now not called anymore.
Made the 2 current methods final and the one that does everything
protected. Made sure that we don't call the tableEvent() all the time
because the current method is used for more calls