Related to SVY-10931 : Table Component -Valuelists
If a string is being set from Rhino always set new value - don't do it
only if previous value was null.

Getting the previous value would not work as expected - only if the
property is not nested in arrays/objects.

It still doesn't work properly if nested because the dataprovider is not
computed correctly currently. (both when a string and a datasource is
set)