For: SVY-10964, SVY-11137, SVY-10932, SVY-11027, SVY-10931
For SVY-10964 it was just a small refactor using what was already added
for other cases - in order to make the code a bit more intuitive and
remove an unfounded assumption in code (that design value is the same as
rhino value for all foundset linked underlying types).

Fixed many problems related to format, valuelist - old (that never
worked properly, such as using a
separate foundset in extra-table with the property types listed below,
or number format on columns with valuelists in extra-table)
and new ones (that the cases above have not yet covered in previous
commits). Basically the way format, dataprovider and valuelist property
types are able to correctly find their dependencies at runtime is
enhanced a lot.

Dependencies can be searched for at runtime as well in current nested
object, any parent level of custom object up to root web component
properties. And they are able to wait for the dependencies to get
propertly set up before using them.

There are also various fixes in the way format and valuelist determine
their dataprovider type and in the way such properties are set from
Rhino.