For: SVY-10964, SVY-11137, SVY-10932, SVY-11027, SVY-10931
When updating some of the unit tests I noticed a problem where if one
would change by reference 'smart' contents in an array, custom object or
component, it would have sent to client a NO-OP instead of the full
value of the new - changed by ref 'smart' value. (smart value = any
value that is change aware and can report that it has no changes)

Added more tests to check for the above problem as well.

+ more helpful error message if getting a column data fails due to wrong
or unavailable column converters.