For: SVY-10964, SVY-11137, SVY-10932, SVY-11027, SVY-10931
As now we removed the native proxies due to the mess they made due to
converting all the time to sablo for any little thing, popup menus
(window_server.js) were not longer working because obj. it is now copied
completely (no reference to initial object remains) in rhino-to-sablo
conversion in Servoy. But we can keep the prototype where the methods
are defined and then it works fine.