For: SVY-10964, SVY-11137, SVY-10932, SVY-11027, SVY-10931
The vl property type could actually get the foundset property's table
even before the actual foundset was there - similar to what format
property type did. So changed that as it is better to have the valuelist
operational as soon as possible - to avoid for example a rhino access to
it before it was initialized because the form's foundset didn't have any
records.