SVY-11392 Error when accessing a form based on an in-memory datasource
always sort the inmem when read from the tablenode in the creation order
(so it is shown in the table editor in that way and it is created in mem
in that way)

If the dataset has columns or types that are not in sync with the
creation order of the definition then don't create a dataset with the
data (which is in a different column order) but try to insert it again
after creation.
Improved the warning messages.