Fixed SVY-11393 : webcomponent's foundset client object should be null if no foundset was set at designtime
Recently there was a change due to SVY-7470 that asks
IFormElementDefaultValueToSabloComponent to write something in the
template depending on other things the type is or is not instead of how
it was before - writing the type.getDefault() (which was null)
automatically.

So now foundset type can write null in the template itself instead of
always writing a value and relying that that toTemplateJSON doesn't get
called for null.