For SVY-11537 : related fields are not updated in table view when data is changed on the same form
Fixed a regression where, when an apply came for a foundset-linked
component, it would set the property on the component while DAL had the
correct record (identified from client sent pkhash) but then DAL would
change to selected record and then the apply/push would get called. (so
potentially on the wrong record).