For SVY-11581 : Time outs in NGClient invoking API when changing values of PK columns
Previous commit contained a fix to update the svyID and an enhancement
to not send multiple unneeded empty updates to the foundset property
type when it is not affected by one column of one row being changed. But
that generated a regression - the client side foundset listener would no
longer notify updates if only the foundset linked properties (DPs,
components) had updates for that row but not the foundset property
itself...

Fixed that regression. If we detect that situation we do send a special
update on the foundset property itself and can collapse multiple
operations of that type or other types that pottentially overwrite each
other into one.

Also made two more improvements that I think cut the message size being
sent to client in my case to half or less:
- don't send unneeded updates for all linked DP properties s in that row
if only one of them changed
- don't send "idForFoundset" all the time for foundset linked properties