For: SVY-11075 Enhance functionality to re-order component's customTypes
It should be clearer now what each item of the config ghost is (caption
in designer as well can be configured in .spec and it hass multiple
fallback levels - for example a table could use header or if header is
not available then dataprovider) and to which droppable property it
belongs (do not put droppable props on top of each other).

This still needs more work as the responsive code works differently in
regards to ghost container positioning.