SVY-12027 java.lang.NullPointerException
some safe check that we don't call getListCellRendererComponent if that
doesn't have a dataadapterlist. it is very weird how that can happen
Because it should be destroyed then (because a dataadapter should be
created when the renderer is made/initialized)
But if it is destroyed it shouldn't get those events anymore

So added some error logs for various situations, because this does fix
the null pointer but not the underlaying cause.