SVY-12473 Bug in Gelders project
enabled the hasForm check again because it did send now a recreateUI
form that was not visible and not send to the server directly already to
the client. Then the next touch form which had the full state was
completely ignored on the client because it was not really changed
anymore (the form url)

because of other fixes that we ask the client what forms are there when
a websocket reconnects this should now work fine.