SVY-12473 Bug in Gelders project
always just have the session id on the url, including the one that are
stored in the endpoints. So we don't need to filter that out in certain
code paths. Like the new behaver that we after a reconnect of the server
the websocket endpoint will ask if there are forms on the client, those
will have the sessionid in the urls.