SVY-12473 Bug in Gelders project
on every start of a websocket endpoint, ask the client which forms are
there. So if there is a hickup in the connection itself (but not a
reload in the browser) then this will fill in the
formsOnClientForThisEndpoint map to be in sync again with the client. So
hasForm is always correct.
Not sure about the boolean "domAttached" that one will not be synced