Fixed SVY-12537: Log messages due to color value mismapping in Sablo?
Actually this was caused due to a limitation of services that didn't
handle default values properly. Now they should handle default values
better. Now, for property types that are only meant to work with
components, not services warning messages will appear in log/developer
console, but these will just get a null as default value.

Also removed some code from fromDesignValue that seemed to handle a
JSONObject but forward it to fromJSON which would then give a
ClassCastException. I think that was there due to some copy/paste.