For SVY-12553: Servoy extra table doesn't work with related valuelistsh
Aggrid was relying on DP with resolve valuelist having an valuelist
value with empty contents in order to do filtering. So I made it so as
long as it doesn't have to send an array of empty valuelist values
through the websocket but just one it will still send it.