For SVY-12553: Servoy extra table doesn't work with related valuelists
Reverted completely the optimization to not send VL at all to client if
DP resolvealuelist is true. AG grid actually did exibit the same
problems as extra-table (if one would use a related valuelist with
real/display values) + some at editing (if it used a foundset that was
not the form's foundset).

So aggrid will also use the new forFoundset added by this case on the
valuelist but it still wants to be able to call filter client-side, even
if the valuelist is record linked and sent 1 time for each row... So
this is needed in order to make editing work correctly for such cells,
without sending all the values for each row to client... just some empty
VLs that you can call filter on.